Department controller & repository removed
